requirements.txt
ipgetter
requests
python-geoip
python-geoip-geolite2
